+++ RiskOfChaos/EffectHandling/ChaosEffectInfo.cs
@@ -56,13 +56,20 @@ namespace RiskOfChaos.EffectHandling
             }
         }
 
-        public readonly float EffectWeightMultiplierPerActivation;
-        public readonly EffectActivationCountMode EffectRepetitionWeightCalculationMode;
+        readonly ConfigEntry<float> _weightReductionPerActivation;
+        readonly float _weightReductionPerActivationDefaultValue;
+
+        public float EffectWeightMultiplierPerActivation => 1f - Mathf.Clamp01(_weightReductionPerActivation?.Value ?? _weightReductionPerActivationDefaultValue);
+
+        readonly ConfigEntry<EffectActivationCountMode> _effectRepetitionCountMode;
+        readonly EffectActivationCountMode _effectRepetitionCountModeDefaultValue;
+
+        public EffectActivationCountMode EffectRepetitionCountMode => _effectRepetitionCountMode?.Value ?? _effectRepetitionCountModeDefaultValue;
 
         public readonly int HardActivationCountCap;
         public readonly bool HasHardActivationCountCap => HardActivationCountCap >= 0;
 
-        public int ActivationCount => ChaosEffectDispatcher.GetEffectActivationCount(EffectIndex, EffectRepetitionWeightCalculationMode);
+        public int ActivationCount => ChaosEffectDispatcher.GetEffectActivationCount(EffectIndex, EffectRepetitionCountMode);
 
         readonly MethodInfo[] _weightMultSelectorMethods;
         public float TotalSelectionWeight
@@ -133,9 +140,6 @@ namespace RiskOfChaos.EffectHandling
                 Log.Error($"attribute target is not a Type ({attribute.target})");
             }
 
-            EffectWeightMultiplierPerActivation = 1f - (attribute.EffectWeightReductionPercentagePerActivation / 100f);
-            EffectRepetitionWeightCalculationMode = attribute.EffectRepetitionWeightCalculationMode;
-
             HardActivationCountCap = attribute.EffectActivationCountHardCap;
 
             ConfigSectionName = "Effect: " + (attribute.ConfigName ?? Language.GetString(NameToken, "en"));
@@ -143,6 +147,12 @@ namespace RiskOfChaos.EffectHandling
             IsEnabledConfig = Main.Instance.Config.Bind(new ConfigDefinition(ConfigSectionName, "Effect Enabled"), true, new ConfigDescription("If the effect should be able to be picked"));
 
             SelectionWeightConfig = Main.Instance.Config.Bind(new ConfigDefinition(ConfigSectionName, "Effect Weight"), attribute.DefaultSelectionWeight, new ConfigDescription("How likely the effect is to be picked, higher value means more likely, lower value means less likely"));
+
+            _weightReductionPerActivationDefaultValue = attribute.EffectWeightReductionPercentagePerActivation / 100f;
+            _weightReductionPerActivation = Main.Instance.Config.Bind(new ConfigDefinition(ConfigSectionName, "Effect Repetition Reduction Percentage"), _weightReductionPerActivationDefaultValue, new ConfigDescription("The percentage reduction to apply to the weight value per activation, setting this to any value above 0 will make the effect less likely to happen several times"));
+
+            _effectRepetitionCountModeDefaultValue = attribute.EffectRepetitionWeightCalculationMode;
+            _effectRepetitionCountMode = Main.Instance.Config.Bind(new ConfigDefinition(ConfigSectionName, "Effect Repetition Count Mode"), _effectRepetitionCountModeDefaultValue, new ConfigDescription($"Controls how the Reduction Percentage will be applied.\n\n{nameof(EffectActivationCountMode.PerStage)}: Only the activations on the current stage are considered, and the weight reduction is reset on stage start.\n\n{nameof(EffectActivationCountMode.PerRun)}: All activations during the current run are considered."));
         }
 
         public readonly void AddRiskOfOptionsEntries()
@@ -156,6 +166,16 @@ namespace RiskOfChaos.EffectHandling
                 min = 0f,
                 max = 2.5f
             }));
+
+            ChaosEffectCatalog.AddEffectConfigOption(new StepSliderOption(_weightReductionPerActivation, new StepSliderConfig
+            {
+                formatString = "-{0:P0}",
+                increment = 0.01f,
+                min = 0f,
+                max = 1f
+            }));
+
+            ChaosEffectCatalog.AddEffectConfigOption(new ChoiceOption(_effectRepetitionCountMode));
         }
 
         public readonly BaseEffect InstantiateEffect(Xoroshiro128Plus effectRNG)